Error 핸들링 & DB에서 사용자 id 를 더 쉽게 조회할 수 있도록 개선

+++ src/components/display-content.vue
@@ -7,8 +7,8 @@
         {{item.title}}
       </v-toolbar-title>
       <v-spacer/>
-      <v-btn @click="article_update" icon><v-icon>mdi-pencil</v-icon></v-btn> 
-      <v-btn @click="remove" icon><v-icon>mdi-delete</v-icon></v-btn> 
+      <v-btn @click="article_update" icon><v-icon>mdi-pencil</v-icon></v-btn>
+      <v-btn @click="remove" icon><v-icon>mdi-delete</v-icon></v-btn>
       <v-btn @click="$emit('close')" icon><v-icon>mdi-close</v-icon></v-btn> <!-- $emit 으로 부모컴포넌트(article-list.vue)에 close 이벤트 전달 -->
     </v-toolbar>
       <!-- 게시물 내용 표시 -->
@@ -42,11 +42,11 @@ import DisplayTime from '@/components/display-time' // @ : src/
 
 export default {
 	components: { DisplayTime },
-	props: ['document','item'],
+	props: ['document', 'item'],
 	data () {
 		return {
 			content: '',
-            ref: this.$firebase.firestore().collection('learning').doc(this.document)
+			ref: this.$firebase.firestore().collection('learning').doc(this.document)
 		}
 	},
 	mounted () {
@@ -59,6 +59,8 @@ export default {
 			this.content = r.data
 		},
 		async article_update () { // 게시물 수정 함수
+			if (!this.$store.state.fireUser) throw Error('관리자만 가능합니다')
+			// 나중에 글 작성자만 삭제할 수 있게끔 수정해야함
 			this.$router.push(
 				{
 					path: this.$route.path + '/article-update',
@@ -67,16 +69,18 @@ export default {
 				}
 			)
 		},
-        async remove() {  // 트랜잭션 처리를 위해 batch 사용
-            const batch = this.$firebase.firestore().batch()
-            batch.update(this.ref, { count : this.$firebase.firestore.FieldValue.increment(-1)})
-            batch.delete(this.ref.collection('articles').doc(this.item.id))
-            await batch.commit()
-            //await this.ref.collection('articles').doc(this.item.id).delete() // Firestore 에서 삭제 (title,시간정보들만 삭제됨)
-            await this.$firebase.storage().ref().child('learning').child(this.document).child(this.item.id + '.html').delete() 
-            // Storage에 저장된 html 파일까지 삭제 완료
-            this.$emit('close')
-        }
+		async remove () { // 트랜잭션 처리를 위해 batch 사용
+			if (!this.$store.state.fireUser) throw Error('관리자만 가능합니다')
+			// 나중에 글 작성자만 삭제할 수 있게끔 수정해야함
+			const batch = this.$firebase.firestore().batch()
+			batch.update(this.ref, { count: this.$firebase.firestore.FieldValue.increment(-1) })
+			batch.delete(this.ref.collection('articles').doc(this.item.id))
+			await batch.commit()
+			// await this.ref.collection('articles').doc(this.item.id).delete() // Firestore 에서 삭제 (title,시간정보들만 삭제됨)
+			await this.$firebase.storage().ref().child('learning').child(this.document).child(this.item.id + '.html').delete()
+			// Storage에 저장된 html 파일까지 삭제 완료
+			this.$emit('close')
+		}
 	}
 
 }
